pyproject.toml
[project]
name = "vidchat"
version = "0.1.0"
description = "AI chat agent with avatar video and audio responses"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai[openai]>=1.0.0",
    "pydantic>=2.0.0",
    "ollama>=0.3.0",
    "pyttsx3>=2.90",
    "pillow>=10.0.0",
    "piper-tts>=1.2.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
# Voice data preparation dependencies
voice-prep = [
    "yt-dlp>=2023.0.0",
    "openai-whisper>=20230314",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "noisereduce>=3.0.0",
    "pydub>=0.25.0",
]

# Voice cloning dependencies
voice-cloning = [
    "coqui-tts>=0.27.0",
]

# Training and experiment tracking dependencies
training = [
    "mlflow>=2.10.0",
    "matplotlib>=3.7.0",
]

# Web interface dependencies
web = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]

# SadTalker avatar dependencies
# Note: torch, torchvision, torchaudio must be installed separately via:
# uv pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# basicsr, gfpgan, facexlib have build issues with Python 3.13+ and are optional (for quality enhancement)
sadtalker = [
    "safetensors>=0.4.0",
    "omegaconf>=2.3.0",
    "einops>=0.7.0",
    "scipy>=1.10.0",
    "scikit-image>=0.21.0",
    "kornia>=0.6.0",
    "yacs>=0.1.8",
    "pyyaml>=6.0.0",
    "dlib>=19.24.0",
    "imageio-ffmpeg>=0.4.0",
    "pydub>=0.25.0",
]

# All optional dependencies
all = [
    "vidchat[voice-prep,web,sadtalker,training,voice-cloning]",
]

[project.scripts]
vidchat = "vidchat.__main__:main"
vidchat-cli = "vidchat.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vidchat"]

[tool.setuptools.packages.find]
where = ["src"]
